Crate hotg_rune_proc_blocks

Structs

ProcBlockDescriptor
A description of everything a particular proc block is capable of.

Tensor
A multidimensional array with copy-on-write semantics.

TensorDescriptor

TransformDescriptor

Enums

Dimension

Dimensions

Traits

HasOutputs

ProcBlock
The base trait that all proc blocks must implement.

Transform
Process some data, transforming it from one form to another.

Derive Macros

ProcBlock
Derive the ProcBlock trait for a particular type.