Struct hotg_rune_proc_blocks::Tensor

pub struct Tensor<T> { /* fields omitted */ }

A multidimensional array with copy-on-write semantics.

Examples

Implementations

impl<T> Tensor<T>

pub fn new_row_major(
elements: Arc<[T]>,
dimensions: Vec<usize, Global>
) -> Tensor<T>

pub fn new_vector(iter: impl IntoIterator<Item = T>) -> Tensor<T>

pub fn single(value: T) -> Tensor<T>

pub fn zeroed(dimensions: Vec<usize, Global>) -> Tensor<T> where
T: Default,

pub fn filled_with<F>(dimensions: Vec<usize, Global>, f: F) -> Tensor<T> where
F: FnMut() -> T,

pub fn shape(&self) -> Shape<'_> where
T: ReflectionType,

pub fn dimensions(&self) -> &[usize]

Get the Tensor’s dimensions.

pub fn rank(&self) -> usize

The number of dimensions this Tensor has.

pub fn elements(&self) -> &[T]

Get an immutable reference to the underlying elements in this Tensor.

pub fn get_elements_mut(&mut self) -> Option<&mut [T]>

Get a mutable reference to the underlying elements in this Tensor.

pub fn view<const RANK: usize>(&self) -> Option<TensorView<'_, T, RANK>>

pub fn view_mut<'a, 'this, const RANK: usize>(
&'this mut self
) -> Option<TensorViewMut<'a, T, RANK>> where
'this: 'a,
T: Clone,

pub fn as_ptr_and_byte_length(&self) -> (*const u8, usize)

#[must_use]
pub fn map<F, Out>(&self, map: F) -> Tensor<Out> where
F: FnMut(&[usize], &T) -> Out,

Create a new Tensor by applying a function to every element in the current tensor.

This is often an expensive operation.

impl<T> Tensor<T> where
T: Clone,

pub fn make_elements_mut(&mut self) -> &mut [T]