Struct hotg_rune_proc_blocks::Tensor

pub struct Tensor<T> { /* private fields */ }

A multidimensional array with copy-on-write semantics.

Implementations

impl<T> Tensor<T>

pub fn new_row_major(
elements: Arc<[T]>,
dimensions: Vec<usize, Global>
) -> Tensor<T>

pub fn new_vector(iter: impl IntoIterator<Item = T>) -> Tensor<T>

pub fn single(value: T) -> Tensor<T>

pub fn zeroed(dimensions: Vec<usize, Global>) -> Tensor<T> where
T: Default,

pub fn filled_with<F>(dimensions: Vec<usize, Global>, f: F) -> Tensor<T> where
F: FnMut() -> T,

pub fn shape(&self) -> Shape<'_> where
T: AsElementType,

pub fn dimensions(&self) -> &[usize]

Get the Tensor’s dimensions.

pub fn rank(&self) -> usize

The number of dimensions this Tensor has.

pub fn elements(&self) -> &[T]

Get an immutable reference to the underlying elements in this Tensor.

pub fn get_elements_mut(&mut self) -> Option<&mut [T]>

Get a mutable reference to the underlying elements in this Tensor.

pub fn slice<const RANK: usize>(
&self,
leading_indices: &[usize]
) -> Option<TensorView<'_, T, RANK>>

Try to get a contiguous sub-slice of this tensor.

Note

Due to the way tensors are laid out in memory, you can only slice off the leading dimensions.

In order to be well-formed, this requires that RANK + leading_indices.len() equals Tensor::rank(), and that each of the indices are within bounds.

Examples

Say you have a [3, 4, 2] tensor, passing in [0, 1] would give you a 1D TensorView that views the 3 elements at [0, 1, ..].

let input = [
    [[0, 1], [2, 3], [4, 5], [6, 7]],
    [[8, 9], [10, 11], [12, 13], [14, 15]],
    [[16, 17], [18, 19], [20, 21], [22, 23]],
];
let tensor: Tensor<i32> = input.into();

let got = tensor.slice::<1>(&[0, 1]).unwrap();

assert_eq!(got.dimensions(), [2]);
assert_eq!(got.elements(), &input[0][1]);

pub fn slice_mut<const RANK: usize>(
&mut self,
leading_indices: &[usize]
) -> Option<TensorViewMut<'_, T, RANK>> where
T: Clone,

A mutable version of Tensor::slice().

pub fn view<const RANK: usize>(&self) -> Option<TensorView<'_, T, RANK>>

Try to reinterpret this tensor as a RANK-D tensor.

pub fn view_mut<'a, 'this, const RANK: usize>(
&'this mut self
) -> Option<TensorViewMut<'a, T, RANK>> where
'this: 'a,
T: Clone,

The mutable version of Tensor::view().

pub fn as_ptr_and_byte_length(&self) -> (*const u8, usize)

pub fn map<F, Out>(&self, map: F) -> Tensor<Out> where
F: FnMut(&[usize], &T) -> Out,

Create a new Tensor by applying a function to every element in the current tensor.

This is often an expensive operation.

pub fn get(&self, indices: &[usize]) -> Option<&T>

Get a reference to the element with this particular index.

pub fn get_mut(&mut self, indices: &[usize]) -> Option<&mut T> where
T: Clone,

Get a mutable reference to the element with this particular index.

impl<T> Tensor<T> where
T: Clone,

pub fn make_elements_mut(&mut self) -> &mut [T]